+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - added `data-parsley-namespace` feature to allow W3C compliant DOM-API (#424)
   - fixed options.error.container element (#425)
   - fixed priority validator error (#426)
+  - added HTML5 tel support (#334)
 
 **1.2.0 (current stable)**